Move CommodityListing creation into its own class
